+++ api/src/main/kotlin/com/inner/circle/api/controller/merchant/MerchantPaymentController.kt
@@ -6,25 +6,30 @@ import com.inner.circle.api.application.dto.PaymentStatusEventType
 import com.inner.circle.api.controller.PaymentForMerchantV1Api
 import com.inner.circle.api.controller.dto.PaymentApproveDto
 import com.inner.circle.api.controller.dto.PaymentResponse
+import com.inner.circle.api.controller.dto.UserCardDto
 import com.inner.circle.api.controller.request.PaymentApproveRequest
 import com.inner.circle.api.controller.request.PaymentClaimRequest
 import com.inner.circle.core.security.MerchantUserDetails
 import com.inner.circle.core.usecase.PaymentClaimUseCase
 import com.inner.circle.core.usecase.SavePaymentApproveUseCase
+import com.inner.circle.core.usecase.UserCardUseCase
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.tags.Tag
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 import org.springframework.security.core.annotation.AuthenticationPrincipal
+import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PathVariable
 import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestBody
+import com.inner.circle.core.service.dto.UserCardDto as CoreUserCardDto
 
 @Tag(name = "Payments - Merchant", description = "상점 고객(SDK) 결제 관련 API")
 @PaymentForMerchantV1Api
 class MerchantPaymentController(
     private val claimUseCase: PaymentClaimUseCase,
     private val savePaymentApproveService: SavePaymentApproveUseCase,
+    private val userCardUseCase: UserCardUseCase,
     private val statusChangedMessageSender: PaymentStatusChangedMessageSender
 ) {
     private val logger: Logger = LoggerFactory.getLogger(MerchantPaymentController::class.java)
@@ -130,5 +135,24 @@ class MerchantPaymentController(
         } catch (e: Exception) {
             logger.error("Error while send ${status.name} Status.", e)
         }
+    }
+
+    @Operation(summary = "모든 카드 조회")
+    @GetMapping("/payments/all/card")
+    fun getAllCard(): PaymentResponse<List<UserCardDto>> {
+        val coreUserCardDtoList = userCardUseCase.findAll()
+        return PaymentResponse.ok(
+            coreUserCardDtoList
+                .map { coreUserCardDto ->
+                    UserCardDto(
+                        id = coreUserCardDto.id,
+                        accountId = coreUserCardDto.accountId,
+                        isRepresentative = coreUserCardDto.isRepresentative,
+                        cardNumber = coreUserCardDto.cardNumber,
+                        expirationPeriod = coreUserCardDto.expirationPeriod,
+                        cvc = coreUserCardDto.cvc
+                    )
+                }.toList()
+        )
     }
 }
