+++ infra/src/main/kotlin/com/inner/circle/infra/adaptor/ConfirmPaymentAdaptor.kt
@@ -20,7 +20,12 @@ internal class ConfirmPaymentAdaptor(
                     request.orderId
                 )
 
-        val userCard = paymentRequest.accountId?.let { userCardRepository.findByAccountId(it) }
+        val userCard = paymentRequest.accountId?.let {
+            userCardRepository.findByAccountIdAndIsRepresentative(
+                it,
+                true
+            )
+        }
 
         return ConfirmPaymentInfraDto(
             orderId = paymentRequest.orderId,
