+++ infra/src/main/kotlin/com/inner/circle/infra/repository/UserCardJpaRepository.kt
@@ -4,5 +4,12 @@ import com.inner.circle.infra.repository.entity.UserCardEntity
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface UserCardJpaRepository : JpaRepository<UserCardEntity, String> {
-    fun findByAccountId(accountId: Long): UserCardEntity?
+    fun findByAccountIdAndIsRepresentative(
+        accountId: Long,
+        isRepresentative: Boolean
+    ): UserCardEntity?
+
+    fun findByAccountId(
+        accountId: Long
+    ): List<UserCardEntity>
 }
