+++ infra/src/main/kotlin/com/inner/circle/infra/repository/UserCardRepository.kt
@@ -3,7 +3,16 @@ package com.inner.circle.infra.repository
 import com.inner.circle.infra.repository.entity.UserCardEntity
 
 interface UserCardRepository {
-    fun findByAccountId(accountId: Long): UserCardEntity?
+    fun findByAccountIdAndIsRepresentative(
+        accountId: Long,
+        isRepresentative: Boolean
+    ): UserCardEntity?
+
+    fun findByAccountId(
+        accountId: Long
+    ): List<UserCardEntity>
+
+    fun findAll(): List<UserCardEntity>
 
     fun save(userCardEntity: UserCardEntity): UserCardEntity?
 }
